requirements.txt
pytesseract==0.3.13
pillow==11.0.0
pypdf==5.1.0
langchain-community==0.3.7
faiss-cpu==1.9.0
langchain-huggingface==0.1.2
sentence-transformers==3.3.0
langchain-groq==0.2.1
Flask==2.2.5
dash==2.18.2
dash-bootstrap-components==1.6.0
python-decouple==3.8